requirements.txt
# Flask core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Forms and validation
WTForms==3.1.1

# HTTP requests for API calls
requests==2.31.0

# RSS feed parsing
feedparser==6.0.11

# Environment variables
python-dotenv==1.0.0

# Date/time utilities
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
Flask-Testing==0.8.1

# WSGI server for production
gunicorn==21.2.0